requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
httpx>=0.27.0
pydantic>=2.9.0
pydantic-settings>=2.6.0
sqlmodel>=0.0.22
openai>=1.55.0
google-generativeai>=0.8.0
python-dotenv>=1.0.0
